fix bordi

+++ include/Hitbox.hpp
@@ -2,7 +2,7 @@
 #pragma once
 
 struct Hitbox {
-    int x, y, width, height;
+    float x, y, width, height;
 };
 
 bool collides(const Hitbox &a, const Hitbox &b);
